2024-06-19 删除无用的模块，修复warning

+++ journalkeeper-utils/src/main/java/io/journalkeeper/utils/parser/EntryParser.java
@@ -45,9 +45,9 @@ public class EntryParser {
 
         byte b = getByte(messageBuffer, byteOffset);
         if (bitValue) {
-            b |= 1 << bitOffset;
+            b |= (byte) (1 << bitOffset);
         } else {
-            b &= ~(1 << bitOffset);
+            b &= (byte) ~(1 << bitOffset);
         }
         setByte(messageBuffer, byteOffset, b);
     }
